refactored a little for better readability
