+++ azure-kusto-data/azure/kusto/data/response.py
@@ -6,164 +6,9 @@ from abc import ABCMeta, abstractmethod
 
 import six
 
-from enum import Enum
 import dateutil.parser
 
-from . import converters
-
-
-class WellKnownDataSet(Enum):
-    """Categorizes data tables according to the role they play in the data set that a Kusto query returns."""
-
-    PrimaryResult = "PrimaryResult"
-    QueryCompletionInformation = "QueryCompletionInformation"
-    TableOfContents = "TableOfContents"
-    QueryProperties = "QueryProperties"
-
-
-class KustoResultRow(object):
-    """Iterator over a Kusto result row."""
-
-    def __init__(self, columns, row):
-        self.columns = columns
-        self.row = row
-
-        # Here we keep converter functions for each type that we need to take special care
-        # (e.g. convert)
-        self.convertion_funcs = {
-            "datetime": converters.to_datetime,
-            "timespan": converters.to_timedelta,
-            "DateTime": converters.to_datetime,
-            "TimeSpan": converters.to_timedelta,
-        }
-
-    @property
-    def columns_count(self):
-        return len(self.columns)
-
-    def __iter__(self):
-        for i in range(self.columns_count):
-            yield self[i]
-
-    def __getitem__(self, key):
-        if type(key) == int:
-            column = self.columns[key]
-            value = self.row[key]
-        else:
-            column = next((column for column in self.columns if column.column_name == key), None)
-            if not column:
-                raise LookupError(key)
-            value = self.row[column.ordinal]
-        if column.column_type in self.convertion_funcs:
-            return self.convertion_funcs[column.column_type](value)
-        return value
-
-    def __len__(self):
-        return self.columns_count
-
-    def __repr__(self):
-        return "KustoResultRow({},{})".format(self.columns, self.row)
-
-
-class KustoResultColumn(object):
-    def __init__(self, json_column, ordianl):
-        self.column_name = json_column["ColumnName"]
-        column_type = json_column.get("ColumnType")
-        self.column_type = column_type or json_column["DataType"]
-        self.ordinal = ordianl
-
-    def __repr__(self):
-        return "KustoResultColumn({},{})".format(
-            json.dumps({"ColumnName": self.column_name, "ColumnType": self.column_type}), self.ordinal
-        )
-
-
-class KustoResultTable(object):
-    """Iterator over a Kusto result table."""
-
-    def __init__(self, json_table):
-        self.table_name = json_table.get("TableName")
-        self.table_id = json_table.get("TableId")
-        self.table_kind = WellKnownDataSet[json_table["TableKind"]] if "TableKind" in json_table else None
-        self.columns = [KustoResultColumn(column, index) for index, column in enumerate(json_table["Columns"])]
-
-        self.rows = json_table["Rows"]
-
-    @property
-    def rows_count(self):
-        return len(self.rows)
-
-    @property
-    def columns_count(self):
-        return len(self.columns)
-
-    def __len__(self):
-        return self.rows_count
-
-    def __iter__(self):
-        for row in self.rows:
-            yield KustoResultRow(self.columns, row)
-
-    def __getitem__(self, key):
-        return KustoResultRow(self.columns, self.rows[key])
-
-    def __str__(self):
-        return "<{table_name} : rows [{rows_count}], cols [{cols_count}]>".format(
-            self.table_name, self.rows_count, self.columns_count
-        )
-
-    # TODO: not sure why this is here...
-    def to_dataframe(self, errors="raise"):
-        import pandas
-
-        kusto_to_dataframe_data_types = {
-            "bool": "bool",
-            "uint8": "int64",
-            "int16": "int64",
-            "uint16": "int64",
-            "int": "int64",
-            "uint": "int64",
-            "long": "int64",
-            "ulong": "int64",
-            "float": "float64",
-            "real": "float64",
-            "decimal": "float64",
-            "string": "object",
-            "datetime": "datetime64[ns]",
-            "guid": "object",
-            "timespan": "timedelta64[ns]",
-            "dynamic": "object",
-            # Support V1
-            "DateTime": "datetime64[ns]",
-            "Int32": "int32",
-            "Int64": "int64",
-            "Double": "float64",
-            "String": "object",
-            "SByte": "object",
-            "Guid": "object",
-            "TimeSpan": "object",
-        }
-
-        """Returns Pandas data frame."""
-        if not self.columns or not self.rows:
-            return pandas.DataFrame()
-
-        frame = pandas.DataFrame(self.rows, columns=[column.column_name for column in self.columns])
-
-        for column in self.columns:
-            col_name = column.column_name
-            col_type = column.column_type
-            if col_type.lower() == "timespan":
-                frame[col_name] = pandas.to_timedelta(
-                    frame[col_name].apply(lambda t: t.replace(".", " days ") if t and "." in t.split(":")[0] else t)
-                )
-            elif col_type.lower() == "dynamic":
-                frame[col_name] = frame[col_name].apply(lambda x: json.loads(x) if x else None)
-            elif col_type in kusto_to_dataframe_data_types:
-                pandas_type = kusto_to_dataframe_data_types[col_type]
-                frame[col_name] = frame[col_name].astype(pandas_type, errors=errors)
-
-        return frame
+from .models import KustoResultColumn, KustoResultRow, KustoResultTable, WellKnownDataSet
 
 
 @six.add_metaclass(ABCMeta)
